flink_ml_service/requirements.txt
apache-flink==1.18.0
pyflink==1.18.0
kafka-python==2.0.2
psycopg2-binary==2.9.7
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3 